SFX code
So, I cleaned up the SFX file names and created two new files (one
header one cpp) for SFX code. While the functions in SFX.cpp will
probably need to be moved around in the project (especially the
Mix_Chunk pointers, as they prob need to be declared somewhere else),
they are properly documented and should be easy to copy/paste to their
respective places. Once these functions are in their proper place, all
thats left is to call Mix_PlayChannel when certain events occur (e.g.
in the same code that changes graphics, we add a line to play the sound
effect).